requirements.txt
pandas>=2.1.0
numpy>=1.26,<2.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.13.0
matplotlib>=3.8.0
plotly>=5.18.0
streamlit>=1.30.0
requests>=2.31.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0
mysql-connector-python>=8.3.0
